Underwatch/requirements.txt
buttplug==0.3.0
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
pip==22.3.1
setuptools==41.2.0
websockets==10.4
